add /allstats, make /delete safer
