+++ Source/ACE.Server/Command/Handlers/DeveloperDatabaseCommands.cs
@@ -1,3 +1,8 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+using Microsoft.EntityFrameworkCore;
 
 using ACE.Database;
 using ACE.Database.Models.Shard;
@@ -5,12 +10,6 @@ using ACE.Entity.Enum;
 using ACE.Server.Command.Handlers.Processors;
 using ACE.Server.Managers;
 using ACE.Server.Network;
-
-using Microsoft.EntityFrameworkCore;
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
 
 namespace ACE.Server.Command.Handlers
 {
